Fix frontend and backend bugs and issues
Fixed the patch/put functionality for user in the backend, in order to include the changes of hashing passwords.
Fixed patching of activity type on backend and frontend.
Fixed overflows on the frontend
Fixed the /Login and /Register page to include the background container and the positioning of the card.
Fixed the background on /Account page.

Contributors:
@jeskanen
@chrtsa